week-5/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.28.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.4.2
pydantic-settings==2.0.3
python-dateutil==2.8.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
email-validator==2.1.0
alembic==1.12.1
psycopg2-binary==2.9.9
argon2-cffi==23.1.0
passlib[argon2]==1.7.4
types-passlib==1.7.4
types-python-jose==3.3.4